Retire old integration test startup/shutdown utils, tidy up a bit
