+++ test/integration/http/misc-test.js
@@ -5,7 +5,7 @@ let tiny = require('tiny-json-http')
 let test = require('tape')
 let sut = join(process.cwd(), 'src')
 let sandbox = require(sut)
-let { url, startupNew: startup, shutdownNew: shutdown, checkHttpResult: checkResult } = require('../../utils')
+let { url, startup, shutdown, checkHttpResult: checkResult } = require('../../utils')
 
 let mock = join(process.cwd(), 'test', 'mock')
 let tmp = join(mock, 'tmp')
