+++ test/integration/lambda-termination-test.js
@@ -5,7 +5,7 @@ let { join } = require('path')
 let { events } = require('../../src')
 let { sync: rm } = require('rimraf')
 let mock = join(process.cwd(), 'test', 'mock')
-let { port, startupNew: startup, shutdownNew: shutdown } = require('../utils')
+let { port, startup, shutdown } = require('../utils')
 let { getPorts } = require(join(process.cwd(), 'src', 'lib', 'ports'))
 let tmp = join(mock, 'tmp')
 let fileThatShouldNotBeWritten = join(tmp, 'do-not-write-me')
