+++ test/integration/plugin-service-test.js
@@ -3,7 +3,7 @@ let { existsSync } = require('fs')
 let test = require('tape')
 let sut = join(process.cwd(), 'src')
 let sandbox = require(sut)
-let { startupNew: startup, shutdownNew: shutdown } = require('../utils')
+let { startup, shutdown } = require('../utils')
 let mock = join(process.cwd(), 'test', 'mock')
 let syncFile = join(mock, 'plugins-sync', 'syncplugin.test')
 let asyncFile = join(mock, 'plugins-async', 'asyncplugin.test')
