+++ test/utils/index.js
@@ -1,13 +1,11 @@
 let checkResult = require('./_check-result')
 let sidechannel = require('./_sidechannel')
 let startupShutdown = require('./_startup-shutdown')
-let startupShutdownNew = require('./_startup-shutdown-new')
 let lib = require('./_lib')
 
 module.exports = {
   ...checkResult,
   ...sidechannel,
   ...startupShutdown,
-  ...startupShutdownNew,
   ...lib,
 }
